# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CONVIVIUM API",
    "version": "0.1.0",
    "description": "Modo edge"
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/v1/catalog/products": {
      "post": {
        "tags": [
          "catálogo"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categoryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "integer",
                    "description": "Precio con impuestos incluidos"
                  },
                  "iepsPct": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 200,
                    "default": 0
                  },
                  "targetPrepSec": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "minimum": 0
                  },
                  "stationIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "minItems": 1
                  },
                  "photoUrl": {
                    "type": "string",
                    "format": "uri",
                    "nullable": true
                  },
                  "active": {
                    "type": "boolean"
                  },
                  "soldOut": {
                    "type": "boolean"
                  },
                  "modifierGroupIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "default": []
                  }
                },
                "required": [
                  "categoryId",
                  "name",
                  "price",
                  "targetPrepSec",
                  "stationIds",
                  "photoUrl",
                  "active",
                  "soldOut"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/catalog/products/{id}": {
      "put": {
        "tags": [
          "catálogo"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categoryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "integer",
                    "description": "Precio con impuestos incluidos"
                  },
                  "iepsPct": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 200,
                    "default": 0
                  },
                  "targetPrepSec": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "minimum": 0
                  },
                  "stationIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "minItems": 1
                  },
                  "photoUrl": {
                    "type": "string",
                    "format": "uri",
                    "nullable": true
                  },
                  "active": {
                    "type": "boolean"
                  },
                  "soldOut": {
                    "type": "boolean"
                  },
                  "modifierGroupIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "default": []
                  }
                },
                "required": [
                  "categoryId",
                  "name",
                  "price",
                  "targetPrepSec",
                  "stationIds",
                  "photoUrl",
                  "active",
                  "soldOut"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/catalog/products/{id}/sold-out": {
      "put": {
        "tags": [
          "catálogo"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "soldOut": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "soldOut"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/catalog/reasons": {
      "get": {
        "tags": [
          "catálogo"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "kind",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ]
}